Create Solitaire
The game should have a number (self.num_piles) of card piles. Each card pile is stored as a CardPile. For example, if self.num_piles is 4, then 4 rows of numbers should be printed out. However, the first row will display the first item only, the others are hidden by an '*'. Thus, you can implement this function by calling self.piles[i].print_all(i) for each pile i. Each line should start with the pile number followed by ": " and the items of the list separated by spaces. An example is shown below:

piles[0] is [6, 5, 4]
piles[1] is [3, 2, 1, 0]
piles[2] is [14, 13, 12, 11]
piles[3] is [10, 9, 8, 7]
The display should be: 
0: 6 * *  
1: 3 2 1 0  
2: 14 13 12 11  
3: 10 9 8 7