Create Move
The method move(self, p1, p2) moves the card from pile p1 to pile p2. There are three types of moves.

Condition 1 (p1 = p2 = 0): move a card from the top of the first pile (front of the list) to the bottom of the first pile (rear of the list). This move is always valid. If the pile is empty, the move does not change anything.

0: 6 2 1
1: 3 2 1 0  

After move(0, 0),

0: 2 1 6
1: 3 2 1 0 

Condition 2 (p1 = 0, p2 > 0): move a card from the top of the first pile (front of the list) to the bottom of any other pile p2 (rear of the list). This move is valid only if pile p1 is not empty. If pile p2 is not empty then the value of the moving card (card1) must be one less than the value of the card (card2) at the bottom of the destination pile by one (i.e., card1 = card2 - 1).

If the move is not valid, it does not change anything.

0: 4 8
1: 7 6 5  

After move(0, 1),

0: 8
1: 7 6 5 4

Condition 3 (p1 > 0, p2 > 0): move the whole pile of cards from pile p1 to the bottom of any other pile p2 (rear of the list). This move is valid only if both piles are not empty and if the value card1 of the card on the top of the moving pile is one less than the value card2 of the card at the bottom of the destination pile (i.e., card1 = card2 - 1). If the move is not valid, it does not change anything.

0: 8
1: 7 6 5  
2: 4 3 2

After move(2, 1),

0: 8
1: 7 6 5 4 3 2
2: